requirements.txt
kafka-python
pyspark
faust-streaming
transformers
streamlit
spacy
sqlalchemy
psycopg2-binary
requests
python-dotenv 
torch
praw